sort imports && add wandb

+++ marl-grid/env/marlgrid/utils/video.py
@@ -1,8 +1,9 @@
+import os
+
+import cv2
 import gym
 import numpy as np
-import os
 import tqdm
-import cv2
 
 
 def export_video(X, outfile, fps=30, rescale_factor=2):
@@ -17,7 +18,7 @@ def export_video(X, outfile, fps=30, rescale_factor=2):
     if isinstance(X, list):
         X = np.stack(X)
 
-    if isinstance(X, np.float) and X.max() < 1:
+    if isinstance(X, float) and X.max() < 1:
         X = (X * 255).astype(np.uint8).clip(0, 255)
 
     if rescale_factor is not None and rescale_factor != 1:
